address comments by @minwoox and @trustin
