+++ brave/brave6/src/main/java/com/linecorp/armeria/server/brave/BraveRpcService.java
@@ -1,7 +1,7 @@
 /*
- * Copyright 2025 LINE Corporation
+ * Copyright 2025 LY Corporation
  *
- * LINE Corporation licenses this file to you under the Apache License,
+ * LY Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
  * with the License. You may obtain a copy of the License at:
  *
@@ -22,13 +22,13 @@ import java.util.function.Function;
 
 import com.linecorp.armeria.common.RpcRequest;
 import com.linecorp.armeria.common.RpcResponse;
+import com.linecorp.armeria.common.annotation.UnstableApi;
 import com.linecorp.armeria.common.brave.RequestContextCurrentTraceContext;
 import com.linecorp.armeria.common.logging.RequestLog;
 import com.linecorp.armeria.server.RpcService;
 import com.linecorp.armeria.server.ServiceRequestContext;
 
 import brave.Span;
-import brave.Tracer;
 import brave.Tracing;
 import brave.rpc.RpcRequestParser;
 import brave.rpc.RpcResponseParser;
@@ -41,17 +41,18 @@ import brave.rpc.RpcTracing;
  * Decorates an {@link RpcService} to trace inbound {@link RpcRequest}s using
  * <a href="https://github.com/openzipkin/brave">Brave</a>.
  */
+@UnstableApi
 public final class BraveRpcService extends AbstractBraveService<RpcServerRequest, RpcServerResponse,
         RpcRequest, RpcResponse> implements RpcService {
 
     private static final RpcRequestParser defaultRequestParser = (request, context, span) -> {
         RpcRequestParser.DEFAULT.parse(request, context, span);
-        ArmeriaRpcServerParser.requestParser().parse(request, context, span);
+        BraveRpcServerParsers.requestParser().parse(request, context, span);
     };
 
     private static final RpcResponseParser defaultResponseParser = (response, context, span) -> {
         RpcResponseParser.DEFAULT.parse(response, context, span);
-        ArmeriaRpcServerParser.responseParser().parse(response, context, span);
+        BraveRpcServerParsers.responseParser().parse(response, context, span);
     };
 
     /**
@@ -74,26 +75,22 @@ public final class BraveRpcService extends AbstractBraveService<RpcServerRequest
         return service -> new BraveRpcService(service, rpcTracing);
     }
 
-    private final Tracer tracer;
     private final RpcServerHandler handler;
-    private final RequestContextCurrentTraceContext currentTraceContext;
 
     private BraveRpcService(RpcService delegate, RpcTracing rpcTracing) {
-        super(delegate);
-        final Tracing tracing = rpcTracing.tracing();
-        tracer = tracing.tracer();
+        super(delegate, rpcTracing.tracing().tracer(),
+              (RequestContextCurrentTraceContext) rpcTracing.tracing().currentTraceContext());
         handler = RpcServerHandler.create(rpcTracing);
-        currentTraceContext = (RequestContextCurrentTraceContext) tracing.currentTraceContext();
     }
 
     @Override
     RpcServerRequest braveRequest(ServiceRequestContext ctx) {
-        return ServiceRequestContextAdapter.asRpcServerRequest(ctx);
+        return RpcServiceRequestContextAdapter.asRpcServerRequest(ctx);
     }
 
     @Override
     RpcServerResponse braveResponse(ServiceRequestContext ctx, RequestLog log, RpcServerRequest braveReq) {
-        return ServiceRequestContextAdapter.asRpcServerResponse(ctx, log, braveReq);
+        return RpcServiceRequestContextAdapter.asRpcServerResponse(ctx, log, braveReq);
     }
 
     @Override
@@ -104,15 +101,5 @@ public final class BraveRpcService extends AbstractBraveService<RpcServerRequest
     @Override
     void handleSend(RpcServerResponse response, Span span) {
         handler.handleSend(response, span);
-    }
-
-    @Override
-    Tracer tracer() {
-        return tracer;
-    }
-
-    @Override
-    RequestContextCurrentTraceContext currentTraceContext() {
-        return currentTraceContext;
     }
 }
