+++ brave/brave6/src/main/java/com/linecorp/armeria/server/brave/ServiceRequestContextAdapter.java
@@ -1,7 +1,7 @@
 /*
- * Copyright 2019 LINE Corporation
+ * Copyright 2025 LY Corporation
  *
- * LINE Corporation licenses this file to you under the Apache License,
+ * LY Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
  * with the License. You may obtain a copy of the License at:
  *
@@ -17,8 +17,6 @@
 package com.linecorp.armeria.server.brave;
 
 import com.linecorp.armeria.common.RequestContext;
-import com.linecorp.armeria.common.RpcRequest;
-import com.linecorp.armeria.common.RpcResponse;
 import com.linecorp.armeria.common.SerializationFormat;
 import com.linecorp.armeria.common.annotation.Nullable;
 import com.linecorp.armeria.common.logging.RequestLog;
@@ -29,8 +27,6 @@ import com.linecorp.armeria.server.ServiceRequestContext;
 
 import brave.Span;
 import brave.http.HttpServerHandler;
-import brave.rpc.RpcServerRequest;
-import brave.rpc.RpcServerResponse;
 
 /**
  * Wraps {@link ServiceRequestContext} in an {@link brave.http.HttpServerRequest}, for use in
@@ -149,105 +145,6 @@ final class ServiceRequestContextAdapter {
     static String serializationFormat(RequestLog requestLog) {
         final SerializationFormat serFmt = requestLog.scheme().serializationFormat();
         return serFmt == SerializationFormat.NONE ? null : serFmt.uriText();
-    }
-
-    static RpcServerRequest asRpcServerRequest(ServiceRequestContext ctx) {
-        return new ArmeriaRpcServerRequest(ctx);
-    }
-
-    private static class ArmeriaRpcServerRequest extends RpcServerRequest {
-
-        private final ServiceRequestContext ctx;
-
-        ArmeriaRpcServerRequest(ServiceRequestContext ctx) {
-            this.ctx = ctx;
-        }
-
-        @Override
-        public long startTimestamp() {
-            return ctx.log().ensureAvailable(RequestLogProperty.REQUEST_START_TIME).requestStartTimeMicros();
-        }
-
-        @Nullable
-        @Override
-        public String method() {
-            final RpcRequest rpcRequest = ctx.rpcRequest();
-            if (rpcRequest == null) {
-                return null;
-            }
-            return rpcRequest.method();
-        }
-
-        @Nullable
-        @Override
-        public String service() {
-            final RpcRequest rpcRequest = ctx.rpcRequest();
-            if (rpcRequest == null) {
-                return null;
-            }
-            return rpcRequest.serviceName();
-        }
-
-        @Nullable
-        @Override
-        protected String propagationField(String keyName) {
-            return ctx.request().headers().get(keyName);
-        }
-
-        @Override
-        public Object unwrap() {
-            return ctx;
-        }
-    }
-
-    static RpcServerResponse asRpcServerResponse(ServiceRequestContext ctx, RequestLog log,
-                                                 RpcServerRequest braveReq) {
-        return new ArmeriaRpcServerResponse(ctx, log, braveReq);
-    }
-
-    private static class ArmeriaRpcServerResponse extends RpcServerResponse {
-
-        private final ServiceRequestContext ctx;
-        private final RequestLog log;
-        private final RpcServerRequest braveReq;
-
-        ArmeriaRpcServerResponse(ServiceRequestContext ctx, RequestLog log, RpcServerRequest braveReq) {
-            this.ctx = ctx;
-            this.log = log;
-            this.braveReq = braveReq;
-        }
-
-        @Override
-        public brave.rpc.RpcRequest request() {
-            return braveReq;
-        }
-
-        @Nullable
-        @Override
-        public String errorCode() {
-            return null;
-        }
-
-        @Nullable
-        @Override
-        public Throwable error() {
-            final Object content = log.responseContent();
-            if (!(content instanceof RpcResponse)) {
-                return null;
-            }
-            final RpcResponse response = (RpcResponse) content;
-            return response.cause();
-        }
-
-        @Override
-        public long finishTimestamp() {
-            return SpanContextUtil.wallTimeMicros(log, log.responseEndTimeNanos());
-        }
-
-        @Override
-        public Object unwrap() {
-            return ctx;
-        }
     }
 
     private ServiceRequestContextAdapter() {}
